apps/finance-ai-agent-demo/backend/requirements.txt
flask>=3.0
flask-socketio>=5.3
flask-cors>=4.0
python-socketio>=5.10
eventlet>=0.35

oracledb>=2.0
langchain-oracledb>=0.2
langchain>=0.2
langchain-community>=0.2
langchain-huggingface>=0.1
langchain-text-splitters>=0.2
sentence-transformers>=3.0

openai>=1.30
tavily-python>=0.3

pdfplumber>=0.10
python-docx>=1.0
PyPDF2>=3.0

numpy<2.0
pandas>=2.0
pydantic>=2.0
python-dotenv>=1.0
tiktoken>=0.7
